chore: colocate core module tests
